+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changes
 
+- Removed `xss` package in favor of `isomorphic-dompurify`
+
 ### Added
 
 ### Fixed